PartyMover: no move InterStepper; poll-based moves
PartyMover used to be an InterStepper, but that wasn't working very well
(even when trying InterStepper's new period_offset), so changed it to
examine 3 polls: current position, one step up, and one step down.  The
step size is from a random generator (Party sets it up with a
[1e-10,0.1) uniform random distribution).  Whichever poll achieves the
highest number of votes is chosen, and the party moves by that step.  If
ties occur, the party randomly chooses one of the best options.