Poll: no longer stores anything, just does polls
Parties no longer use history but current polling to choose a position,
so got rid of Poll's history.  Instead it now can conduct (and return) a
poll for current positions, or current positions for other parties and a
hypothetical position for one party.

Ties in number of voters are now randomized: voters choose a closest
party randomly with equal probability of choosing each.